feat: add crud endpoints

+++ src/database/post_model.go
@@ -11,9 +11,9 @@ type Post struct {
 	UpdatedAt time.Time      `gorm:"not null" json:"updatedAt"`
 	DeletedAt gorm.DeletedAt `gorm:"index" json:"deletedAt"`
 
-	Title      string  `gorm:"not null" json:"title"`
-	Content    string  `gorm:"not null" json:"content"`
-	TimeToRead float64 `gorm:"not null" json:"timeToRead"`
+	Title      string `gorm:"not null" json:"title"`
+	Content    string `gorm:"not null" json:"content"`
+	TimeToRead uint64 `gorm:"not null" json:"timeToRead"`
 
 	UserID uint64 `gorm:"index;not null" json:"userID"`
 }
